Adicionando métodos para atualizar e excluir utilidades

+++ src/api/controllers/feature-controller.ts
@@ -1,14 +1,17 @@
 import { FastifyReply, FastifyRequest } from 'fastify';
 import { featureServices } from '../services/features-services';
-import { Feature } from '@prisma/client';
+import { Feature, User } from '@prisma/client';
 import { ERR_FEATURE_ALREADY_EXISTS, ERR_FEATURE_TYPE_ERROR } from '../errors/feature-errors';
 import { getPagination } from '../helpers/get-pagination';
 import { FeatureModel } from '../../../prisma/models';
 import { getIssuesZod } from '../helpers/get-issues-zod';
 import { CustomError } from '../helpers/custom-error';
+import { ERR_PERMISSION_DENIED } from '../errors/permission-erros';
 
 const featureController = {
   createOneFeature: async (req: FastifyRequest, res: FastifyReply) => {
+    const user = req.user as User;
+    if (user.role !== 'Admin') throw ERR_PERMISSION_DENIED;
     const feature = req.body as Feature;
     if (!['Property', 'Building', 'Community'].includes(feature.type)) throw ERR_FEATURE_TYPE_ERROR;
     try {
@@ -25,6 +28,8 @@ const featureController = {
   },
 
   findAll: async (req: FastifyRequest, res: FastifyReply) => {
+    const user = req.user as User;
+    if (user.role !== 'Admin') throw ERR_PERMISSION_DENIED;
     const { page, per_page } = req.query as { page: string; per_page: string };
     const { page_number, per_page_number, skip } = getPagination(page, per_page);
     try {
@@ -36,6 +41,8 @@ const featureController = {
   },
 
   updateOneFeature: async (req: FastifyRequest, res: FastifyReply) => {
+    const user = req.user as User;
+    if (user.role !== 'Admin') throw ERR_PERMISSION_DENIED;
     const { id } = req.params as { id: string };
     const body = req.body as Feature;
 
@@ -55,6 +62,8 @@ const featureController = {
   },
 
   deleteOneFeature: async (req: FastifyRequest, res: FastifyReply) => {
+    const user = req.user as User;
+    if (user.role !== 'Admin') throw ERR_PERMISSION_DENIED;
     const { id } = req.params as { id: string };
     try {
       const result = await featureServices.deleteOneFeature(Number(id));
